requirements.txt
biopython==1.83
openai>=1.14
pandas>=2.2
jinja2>=3.1